crazy? i was crazy once. they locked me in a room. a rubber room. a rubber room with rats. i hate rats. they make me crazy.

+++ src/kernel/vga.h
@@ -14,5 +14,6 @@ bool kprintf(const char *format, ...);
 bool klog(const char *format, ...);
 
 void putpixel(uint8_t color, size_t x, size_t y);
-void putpixelmatrix(uint16_t begin_x, uint16_t begin_y, uint16_t size_x, uint16_t size_y, uint8_t nowrite_byte, uint8_t *matrix);
+void putpixelmatrix(int begin_x, int begin_y, uint16_t size_x, uint16_t size_y, uint8_t nowrite_byte, uint8_t *matrix);
+// positions can be negative in some cases, we cant leave X and Y as uints as a result.
 #endif
